changed some styles, added new components and some fixes
