--- src/App.scss
@@ -1,6 +1,0 @@
-.wrapper {
-  width: 100%;
-  max-width: 600px;
-  height: 100%;
-  margin: 0 auto;
-}
