+++ src/components/header/Header.module.scss
@@ -5,24 +5,15 @@
   line-height: 1.5em;
   padding-bottom: 15px;
   position: relative;
+  text-align: center;
   &:before {
     content: "";
     position: absolute;
     left: 0;
-    bottom: 0;
-    height: 8px;
-    width: 55px;
+    bottom: 10%;
+    height: 5px;
+    width: 100%;
     background-color: #111;
     z-index: 1;
-  }
-  &:after {
-    content: "";
-    position: absolute;
-    left: 0;
-    bottom: 3px;
-    height: 2px;
-    width: 95%;
-    max-width: 255px;
-    background-color: #333;
   }
 }
