+++ src/components/main/Main.module.scss
@@ -1,0 +1,4 @@
+.main {
+  border: 1px solid #eeeef2;
+  border-radius: 16px;
+}
