+++ src/components/main/purhcases-list/PurchasesList.module.scss
@@ -1,0 +1,3 @@
+.purchasesList {
+  width: 100%;
+}
