+++ src/components/main/purhcases-list/purchases-list-item/PurchasesListItem.module.scss
@@ -1,0 +1,8 @@
+.listItem {
+  display: flex;
+  justify-content: space-between;
+  padding: 10px 25px;
+  &:not(:last-child) {
+    border-bottom: 1px solid #eeeef2;
+  }
+}
