+++ src/components/main/selected-block/SelectedBlock.module.scss
@@ -1,6 +1,6 @@
 .selectBlock {
   display: flex;
-  margin: 10px 0;
+  margin: 10px 10px;
 }
 
 .selectItem {
