--- src/components/purchases-list-header/PurchasesListHeader.tsx
@@ -1,16 +1,0 @@
-import styles from './PurchasesListHeader.module.scss';
-
-const PurchasesListHeader = () => {
-  return (
-    <div>
-      <ul className={styles.list}>
-        <li className={styles.list__item}>№</li>
-        <li className={styles.list__item}>Name</li>
-        <li className={styles.list__item}>Spent</li>
-        <li className={styles.list__item}>Date</li>
-      </ul>
-    </div>
-  );
-};
-
-export default PurchasesListHeader;
