--- src/components/purchases-list-item/PurchasesListItem.module.scss
@@ -1,6 +1,0 @@
-.root {
-  display: flex;
-  justify-content: space-between;
-  width: 500px;
-  margin-bottom: 15px;
-}
